requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.16.1
sqlparse==0.5.5
ruff==0.15.6
